Data Preprocessing
Ingest WN18 Entities and Relations into Relational Datastore
