+++ fixtures/WN18/entity.py
@@ -4,7 +4,6 @@ import sys
 import re
 import copy
 import logging
-from pprint import pprint
 
 # 3rd party
 import psycopg2
@@ -30,8 +29,8 @@ def get_path(dirname, filename=None):
         Returns:
             filepath (str): Full file file path.
         """
-    util_dirname, _ =  os.path.split(__file__)
-    project, _ = os.path.split(util_dirname)
+    fixtures, _ = os.path.split(os.path.dirname(__file__))
+    project = os.path.dirname(fixtures)
     path = os.path.join(project, dirname, filename) if filename else os.path.join(project, dirname)
 
     return path
@@ -56,21 +55,19 @@ def get_connection(user, password, host, port, database):
     return connection
 
 
-def insert_record(record, tablename, cursor, connection):
-
-    print(f'final record: {record}')
+def insert_record(record, tablename, connection):
     columns = record.keys()
     logger.debug(f'columns: {columns}')
     values = record.values()
     logger.debug(f'values: {values}')
     values = list(map(lambda x: Json(x) if isinstance(x, dict) else x, values))
 
-    values = tuple(values)
+    cursor = connection.cursor()
     insert_statement = 'INSERT INTO %s (%s) VALUES %s'
-    logger.debug(f"cursor.mogrify: {cursor.mogrify(insert_statement, (AsIs(tablename), AsIs(','.join(columns)), values))}")
+    logger.debug(f"cursor.mogrify: {cursor.mogrify(insert_statement, (AsIs(tablename), AsIs(','.join(columns)), tuple(values)))}")
 
     try:
-        cursor.execute(insert_statement, (AsIs(tablename), AsIs(','.join(columns)), values))
+        cursor.execute(insert_statement, (AsIs(tablename), AsIs(','.join(columns)), tuple(values)))
     except Exception as e:
         logger.error(f'Could not insert into {tablename}, {e}')
 
@@ -79,36 +76,36 @@ def insert_record(record, tablename, cursor, connection):
     logger.debug(f'{count} Record inserted successfully into {tablename} table')
 
 
-def insert_records(records):
-    with get_connection('scientist',
-                        '*************',
-                        '127.0.0.1',
-                        '5432',
-                        'tensor_factorisation') as connection:
-
-        tablename = 'wn18_relation'
-        cursor = connection.cursor()
-
+def insert_records(records, tablename, connection):
+    with connection as con:
         for record in records:
-            insert_record(record, tablename, cursor, connection)
-
-        count = cursor.rowcount
-        logger.debug(f'{count} Record inserted successfully into {tablename} table')
+            insert_record(record, tablename, con)
 
 
-def get_records(relationfile):
-    with open(relationfile, 'r') as entityfile:
+def get_records(entityfile):
+    with open(entityfile, 'r') as entityfile:
         records = []
         record = {}
 
         for line in entityfile:
-            _, relation, _ = line.strip().split('\t')
-            logger.debug(f'relation: {relation}')
+            synset_id, intelligible_name, definition = line.strip().split('\t')
+            logger.debug(f'synset_id: {synset_id}, intelligible_name: {intelligible_name}, defintion: {definition}')
 
-            doc = relation.replace('_', ' ').strip()
+            pattern = re.compile(r'^__([a-zA-Z0-9\'\._/-]*)_([A-Z]{2})_([0-9])')
+            doc = pattern.search(intelligible_name).group(1).replace('_', ' ')
             logger.debug(f'doc: {doc}')
+            POS_tag = pattern.search(intelligible_name).group(2)
+            logger.debug(f'POS_tag: {POS_tag}')
+            sense_index = pattern.search(intelligible_name).group(3)
+            logger.debug(f'sense_index: {sense_index}')
 
+            definition = definition.replace('_', ' ')
+
+            record['synset_id'] = synset_id
             record['doc'] = doc
+            record['POS_tag'] = POS_tag
+            record['sense_index'] = int(sense_index)
+            record['definition'] = definition
 
             logger.debug(f'record: {record}')
             records.append(copy.copy(record))
@@ -119,29 +116,18 @@ def get_records(relationfile):
 
 
 def main():
+    connection = get_connection('scientist',
+                                '*********',
+                                '127.0.0.1',
+                                '5432',
+                                'wn18')
 
-    relationfile = get_path('data/WN18')
-    logger.debug(f'relationfile: {relationfile}')
+    tablename = 'entity'
 
-    dirname, _, _, _, _ = list(os.walk(relationfile))
-    _, _, filenames = dirname
-
-    records = []
-
-    for filename in filenames:
-        if filename == 'train.txt':
-            logger.debug(f'filename: {filename}')
-            filename = get_path('data/WN18', filename)
-            records_train = [{'doc': value} for value in set([relation['doc'] for relation in get_records(filename)])]
-            logger.debug(f'records_train: {records_train}')
-
-            number_of_records = len(records_train)
-            logger.debug(f'number of relations so far: {number_of_records}')
-            records.extend(records_train)
-            if number_of_records == 18:
-                break
-
-    insert_records(records)
+    entityfile = get_path('data/WN18', 'wordnet-mlj12-definitions.txt')
+    logger.debug(f'entityfile: {entityfile}')
+    records = get_records(entityfile)
+    insert_records(records, tablename, connection)
 
 
 if __name__ == '__main__':
